requirements.txt
manga_ocr==0.1.10
Pillow==9.5.0
